requirements.txt
pytelegrambotapi
openai>=1.0.0
python-dotenv
